refactor(server): extract code-graph tools into graph_tools.py (P1 of the Loci review)
server.py is the review's #1 finding: an 8,953-line, 176-symbol, most-coupled
monolith holding 60 tools + 115 helpers. Start the domain split with the cleanest,
lowest-risk slice — the 11 code<->memory graph tools, which depend only on a
_get_kuzu accessor (no coupling to the investigation/memory internals).

- New graph_tools.py: the 11 tools (code_graph_ingest/query, code_memory_map/relink,
  symbol_impact, impact_report, finding_code_context, investigation_code_briefing,
  subsystem_report, related_investigations_via_code, dead_code_candidates) as plain
  functions + register(mcp, get_kuzu) that injects the dep and registers each on the
  shared FastMCP instance (FastMCP reads signature+docstring, so tools are identical).
- server.py calls graph_tools.register(mcp, _get_kuzu) and re-exports the callables
  so `server.<tool>()` keeps working for in-process callers/tests.

server.py 8953 -> 8588 lines. All 11 tools still register; full suite 269 passed.
Establishes the pattern; the investigation_*/memory_* domains are follow-ups.
